Fix bug in delete order controller

+++ app/controllers/orders_controller.rb
@@ -35,7 +35,7 @@ class OrdersController < ApplicationController
 
   def destroy
     # @order = Order.find(params[:id])
-    Order.delete_all
+    Order.where(["user_id = ?", current_user.id]).delete_all
     # redirect_to orders_path, notice: "The order was successfully deleted"
     redirect_to new_order_path, notice: "The order was successfully cleared"    
   end
